requirements.txt
python-dotenv==1.0.0
langchain==0.1.17
langchain-groq==0.1.3
pinecone-client==4.0.0
crewai==0.28.8